test: --probe-steps single-step gates, Metal backend name, PIXAL3D_WEB_PORT, wasm32 memory-sum fix
trellis-test-pixal3d-slat-sample --probe-steps k1,k2,..: integrate sampler step k once from
the f32 reference's x_step{k-1} and score against x_step{k} (f32 / cuda / cuda_nofa), then
exit -- the development regression for the WebGPU flow stages (one forward instead of 12).
--backend metal now accepts ggml-metal's 'MTL0'. web/texture/run_playwright.js honors
PIXAL3D_WEB_PORT. pixal3d_texture_run sums its device-resident total in double (the 4790 MB
overflowed 32-bit size_t under wasm32 and printed 693.9 MB).

+++ src/pixal3d_wasm.cpp
@@ -465,7 +465,7 @@ int run_texture_impl(const string& dinov3_gguf, const string& naf_gguf, const st
         ggml_backend_name(mf.backend), mf.total_bytes() / 1048576.0, run->alloc_bytes() / 1048576.0,
         (cond_g.size() + proj.size()) * 4 / 1048576.0, (neg_cond.size() + neg_proj.size()) * 4 / 1048576.0,
         (long long)Cin, sample.size() * 4 / 1048576.0, concat_cond.size() * 4 / 1048576.0, (long long)(2 * Cin), (double)2 * Cin * N * 4 / 1048576.0,
-        (mf.total_bytes() + run->alloc_bytes() + (cond_g.size() + proj.size() + neg_cond.size() + neg_proj.size() + sample.size() + 2 * Cin * N) * 4) / 1048576.0);
+        ((double)mf.total_bytes() + (double)run->alloc_bytes() + (double)(cond_g.size() + proj.size() + neg_cond.size() + neg_proj.size() + sample.size() + 2 * Cin * N) * 4) / 1048576.0);   // summed in double: the byte total exceeds 32-bit size_t under wasm32
     int n_fwd = 0; double fwd_ms = 0;
     vector<float> x64((size_t)2 * Cin * N);
     FlowFwdProj fwd = [&](const vector<float>& x, float ts, const float* c, const float* pj) {
